Add option for staging all files during commit
This update introduces an option to stage all files automatically during the commit process. The change includes additions to the commit command to provide a `--addAll` flag for users. Now, if this flag is used, all changes in the current working directory will be staged before committing.

- Added a new boolean parameter `addAllChanges` in the commit function
- Modified the CLI commit command to include an `--addAll` option
- Renamed `addAllChanges` method to `stageAllFiles` for clarity

Release-Note: Add option to stage all files with `--addAll` flag in commit command

Signed-off-by: Ian Skelskey <ianskelskey@gmail.com>